refactor: modularize apts/utils/planetary.py into a package
Split the 1,142-line "God Object" apts/utils/planetary.py into a modular
package structure under apts/utils/planetary/.

- base.py: Core object retrieval and naming normalization.
- constants.py: Physical constants and naming dictionaries.
- moon.py: Moon-specific calculations.
- physical.py: General physical and apparent property calculations.
- planets.py: Specialized models for major planets.
- sun.py: Specialized models for the Sun.
- __init__.py: Re-exports the public API for backward compatibility.

Updated tests to reflect the new internal structure where necessary.
Verified with the unit test suite.
